refactor: FatePolicyFinder 도메인 단으로 이동
