+++ src/main/java/wooteco/subway/path/domain/fare/FarePolicyFinder.java
@@ -1,9 +1,10 @@
-package wooteco.subway.path.application;
+package wooteco.subway.path.domain.fare;
 
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import wooteco.subway.path.domain.fare.*;
 
-@Service
+@Component
 public class FarePolicyFinder {
     public static final int BASIC_FARE = 1250;
     public static final int OVER_FARE = 2050;
